Remove BigQueryV3SourceBase
This change removes the BigQueryV3SourceBase class and modifies the
BigQueryV3TableSource class to inherit directly from BoundedSource.
The intermediate class serves no value, and it doesn't make sense
for BigQueryV3TableSource to inherit from BigQuerySourceBase since
it uses almost none of the common code there for managing jobs etc.
